implement LibXC interface using CFFI ABI mode
